+++ pkpdapp/pkpdapp/api/serializers/dataset.py
@@ -3,260 +3,15 @@
 # is released under the BSD 3-clause license. See accompanying LICENSE.md for
 # copyright notice and full license details.
 #
+
 from rest_framework import serializers
-from pkpdapp.models import (
-    Dataset, BiomarkerType, Subject, Protocol, Project,
-    PharmacokineticModel, PharmacodynamicModel,
-    DosedPharmacokineticModel, PkpdModel,
-    Profile, Dose, Unit, Biomarker, Compound, Variable,
-    ProjectAccess,
-)
-from django.contrib.auth.models import User
-from pkpdapp.models.mechanistic_model import MyokitModelMixin
-from myokit.formats.sbml import SBMLParsingError
 import codecs
 import pandas as pd
-
-
-class ValidSbml:
-    def __call__(self, value):
-        try:
-            MyokitModelMixin.parse_sbml_string(value)
-        except SBMLParsingError as e:
-            raise serializers.ValidationError(e)
-
-
-class DoseSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Dose
-        fields = '__all__'
-
-
-class AuceSerializer(serializers.Serializer):
-    name = serializers.CharField()
-    fit_type = serializers.CharField()
-    subject_ids = serializers.ListField(
-        child=serializers.IntegerField()
-    )
-    concentrations = serializers.ListField(
-        child=serializers.FloatField()
-    )
-    auce = serializers.ListField(
-        child=serializers.FloatField()
-    )
-
-    x = serializers.ListField(
-        child=serializers.FloatField()
-    )
-    y = serializers.ListField(
-        child=serializers.FloatField()
-    )
-    y_upper = serializers.ListField(
-        child=serializers.FloatField()
-    )
-    y_lower = serializers.ListField(
-        child=serializers.FloatField()
-    )
-    fit_EC50 = serializers.FloatField()
-    sigma_EC50 = serializers.FloatField()
-    fit_top = serializers.FloatField()
-    sigma_top = serializers.FloatField()
-    fit_bottom = serializers.FloatField()
-    sigma_bottom = serializers.FloatField()
-
-
-class NcaSerializer(serializers.Serializer):
-    times = serializers.ListField(
-        child=serializers.FloatField()
-    )
-    concentrations = serializers.ListField(
-        child=serializers.FloatField()
-    )
-    dose_amount = serializers.FloatField()
-    administration_route = serializers.CharField()
-    c_0 = serializers.FloatField()
-    auc_0_last = serializers.FloatField()
-    aumc_0_last = serializers.FloatField()
-    lambda_z = serializers.FloatField()
-    r2 = serializers.FloatField()
-    num_points = serializers.IntegerField(min_value=0)
-    auc_infinity = serializers.FloatField()
-    auc_infinity_dose = serializers.FloatField()
-    auc_extrap_percent = serializers.FloatField()
-    cl = serializers.FloatField()
-    c_max = serializers.FloatField()
-    t_max = serializers.FloatField()
-    c_max_dose = serializers.FloatField()
-    aumc = serializers.FloatField()
-    aumc_extrap_percent = serializers.FloatField()
-    mrt = serializers.FloatField()
-    tlast = serializers.FloatField()
-    t_half = serializers.FloatField()
-    v_ss = serializers.FloatField()
-    v_z = serializers.FloatField()
-
-
-class ProtocolSerializer(serializers.ModelSerializer):
-    doses = DoseSerializer(
-        many=True, read_only=True
-    )
-    dose_ids = serializers.PrimaryKeyRelatedField(
-        queryset=Dose.objects.all(),
-        source='doses',
-        many=True, write_only=True,
-    )
-    dosed_pk_models = serializers.PrimaryKeyRelatedField(
-        many=True, read_only=True
-    )
-
-    class Meta:
-        model = Protocol
-        fields = '__all__'
-
-
-class PharmacokineticSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = PharmacokineticModel
-        fields = '__all__'
-
-
-class ComponentSerializerField(serializers.Field):
-
-    VALUE_MAP = {
-        'M': 'Male',
-        'F': 'Female'
-    }
-
-    def to_representation(self, obj):
-        return self.VALUE_MAP[obj]
-
-    def to_internal_value(self, data):
-        return {k: v for v, k in self.VALUE_MAP.items()}[data]
-
-
-def _serialize_component(model, component, myokit_model):
-
-    states = [
-        v.pk
-        for v in model.variables.filter(state=True)
-        if v.qname.startswith(component.name())
-    ]
-
-    variables = [
-        v.pk
-        for v in model.variables.filter(constant=True)
-        if v.qname.startswith(component.name())
-    ]
-
-    outputs = [
-        v.pk
-        for v in model.variables.filter(constant=False)
-        if v.qname.startswith(component.name())
-    ]
-
-    equations = [
-        MyokitModelMixin._serialise_equation(e)
-        for e in component.equations(bound=False, const=False)
-    ]
-    return {
-        'name': component.name(),
-        'states': states,
-        'variables': variables,
-        'outputs': outputs,
-        'equations': equations,
-    }
-
-
-class DosedPharmacokineticSerializer(serializers.ModelSerializer):
-    components = serializers.SerializerMethodField('get_components')
-    variables = serializers.PrimaryKeyRelatedField(
-        many=True, read_only=True
-    )
-
-    def get_components(self, m):
-        model = m.get_myokit_model()
-        return [
-            _serialize_component(m, c, model)
-            for c in model.components(sort=True)
-        ]
-
-    class Meta:
-        model = DosedPharmacokineticModel
-        fields = '__all__'
-
-
-class PharmacodynamicSerializer(serializers.ModelSerializer):
-    components = serializers.SerializerMethodField('get_components')
-    variables = serializers.PrimaryKeyRelatedField(
-        many=True, read_only=True
-    )
-
-    def get_components(self, m):
-        model = m.get_myokit_model()
-        return [
-            _serialize_component(m, c, model)
-            for c in model.components(sort=True)
-        ]
-
-    class Meta:
-        model = PharmacodynamicModel
-        exclude = ['sbml']
-
-
-class PharmacodynamicSbmlSerializer(serializers.ModelSerializer):
-    sbml = serializers.CharField(validators=[ValidSbml()])
-
-    class Meta:
-        model = PharmacodynamicModel
-        fields = ['sbml']
-
-
-class PkpdSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = PkpdModel
-        fields = '__all__'
-
-
-# all serializers that get used by the dataset serializer
-
-class UnitSerializer(serializers.ModelSerializer):
-    compatible_units = \
-        serializers.SerializerMethodField('get_compatible_units')
-
-    class Meta:
-        model = Unit
-        fields = '__all__'
-
-    def get_compatible_units(self, unit):
-        return [
-            {
-                'id': u.id,
-                'symbol': u.symbol,
-            } for u in unit.get_compatible_units()
-        ]
-
-
-class VariableSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Variable
-        fields = '__all__'
-
-
-class BiomarkerTypeSerializer(serializers.ModelSerializer):
-    data = serializers.SerializerMethodField('get_data')
-
-    class Meta:
-        model = BiomarkerType
-        fields = '__all__'
-
-    def get_data(self, bt):
-        return bt.as_pandas().to_dict(orient='list')
-
-
-class SubjectSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Subject
-        fields = '__all__'
+from pkpdapp.models import (
+    Dataset, BiomarkerType, Protocol, Subject,
+    Unit, Compound, Biomarker
+)
+from pkpdapp.api.serializers import ProtocolSerializer
 
 
 class DatasetSerializer(serializers.ModelSerializer):
@@ -486,40 +241,3 @@ class DatasetCsvSerializer(serializers.ModelSerializer):
                     protocol=protocol,
                 )
         return instance
-
-
-class ProfileSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Profile
-        fields = '__all__'
-
-
-class UserSerializer(serializers.ModelSerializer):
-    profile = ProfileSerializer(read_only=True)
-    project_set = serializers.PrimaryKeyRelatedField(
-        many=True, read_only=True,
-    )
-
-    class Meta:
-        model = User
-        fields = '__all__'
-        fields = (
-            'id', 'username', 'first_name',
-            'last_name', 'email', 'profile', 'project_set'
-        )
-
-
-class ProjectAccessSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ProjectAccess
-        fields = '__all__'
-
-
-class ProjectSerializer(serializers.ModelSerializer):
-    user_access = ProjectAccessSerializer(
-        source='projectaccess_set', many=True, read_only=True
-    )
-
-    class Meta:
-        model = Project
-        fields = '__all__'
