+++ pkpdapp/pkpdapp/api/serializers/models.py
@@ -1,0 +1,101 @@
+#
+# This file is part of PKPDApp (https://github.com/pkpdapp-team/pkpdapp) which
+# is released under the BSD 3-clause license. See accompanying LICENSE.md for
+# copyright notice and full license details.
+#
+from rest_framework import serializers
+from pkpdapp.models import (
+    PharmacokineticModel, MyokitModelMixin,
+    DosedPharmacokineticModel, PkpdModel,
+    PharmacodynamicModel
+)
+
+class PkpdSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = PkpdModel
+        fields = '__all__'
+
+
+class PharmacokineticSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = PharmacokineticModel
+        fields = '__all__'
+
+
+def _serialize_component(model, component, myokit_model):
+
+    states = [
+        v.pk
+        for v in model.variables.filter(state=True)
+        if v.qname.startswith(component.name())
+    ]
+
+    variables = [
+        v.pk
+        for v in model.variables.filter(constant=True)
+        if v.qname.startswith(component.name())
+    ]
+
+    outputs = [
+        v.pk
+        for v in model.variables.filter(constant=False)
+        if v.qname.startswith(component.name())
+    ]
+
+    equations = [
+        MyokitModelMixin._serialise_equation(e)
+        for e in component.equations(bound=False, const=False)
+    ]
+    return {
+        'name': component.name(),
+        'states': states,
+        'variables': variables,
+        'outputs': outputs,
+        'equations': equations,
+    }
+
+
+class DosedPharmacokineticSerializer(serializers.ModelSerializer):
+    components = serializers.SerializerMethodField('get_components')
+    variables = serializers.PrimaryKeyRelatedField(
+        many=True, read_only=True
+    )
+
+    def get_components(self, m):
+        model = m.get_myokit_model()
+        return [
+            _serialize_component(m, c, model)
+            for c in model.components(sort=True)
+        ]
+
+    class Meta:
+        model = DosedPharmacokineticModel
+        fields = '__all__'
+
+
+class PharmacodynamicSerializer(serializers.ModelSerializer):
+    components = serializers.SerializerMethodField('get_components')
+    variables = serializers.PrimaryKeyRelatedField(
+        many=True, read_only=True
+    )
+
+    def get_components(self, m):
+        model = m.get_myokit_model()
+        return [
+            _serialize_component(m, c, model)
+            for c in model.components(sort=True)
+        ]
+
+    class Meta:
+        model = PharmacodynamicModel
+        exclude = ['sbml']
+
+
+class PharmacodynamicSbmlSerializer(serializers.ModelSerializer):
+    sbml = serializers.CharField(validators=[ValidSbml()])
+
+    class Meta:
+        model = PharmacodynamicModel
+        fields = ['sbml']
+
+
