+++ pkpdapp/pkpdapp/api/serializers/variables.py
@@ -1,0 +1,19 @@
+#
+# This file is part of PKPDApp (https://github.com/pkpdapp-team/pkpdapp) which
+# is released under the BSD 3-clause license. See accompanying LICENSE.md for
+# copyright notice and full license details.
+#
+from rest_framework import serializers
+from pkpdapp.models import (
+    Variable, StoredVariable
+)
+
+class VariableSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Variable
+        fields = '__all__'
+
+class StoredVariableSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = StoredVariable
+        fields = '__all__'
